lists: support smtp ingestion
To avoid issues such as [1] creeping into Production, this improves the
lists setup so that they can actually receive new messages, which
properly exercises the *nominal* inbound flow (bouncing does not work).

The new script (disclaimer: I used ChatGPT to write it) can also be used
to interact with todos.

[1] https://git.sr.ht/~sircmpwn/lists.sr.ht/commit/a74f0d77

Signed-off-by: Simon Martin <simon@nasilyan.com>